List booking and vehicles

+++ Marathon/Marathon.Domain/Constants/MaintenanceConstants.cs
@@ -7,7 +7,7 @@ namespace Marathon.Domain.Constants
 {
     public static class MaintenanceConstants
     {
-        public static readonly int MileageBetweenChecks = 10000;
+        public static readonly decimal MileageBetweenChecks = 10000;
         public static readonly TimeSpan DurationBetweenChecks = new TimeSpan(365, 0, 0, 0);
     }
 }
